[#163] Refactor food request view to use separate modules
- Moved the FoodRequestViewSet class to a separate module `food_request.py` to improve code organization and maintainability
- This change separates the view logic from the main `__init__.py` file, making the codebase more modular and easier to navigate
- The core functionality of the FoodRequestViewSet remains unchanged, but the restructuring will facilitate future additions or modifications to the food request handling
